requirements.txt
streamlit
langchain-google-genai
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu